--- src/main/java/racingcar/model/Car.java
@@ -1,5 +1,0 @@
-package racingcar.model;
-
-public interface Car {
-
-}
