+++ src/main/java/racingcar/model/car/Car.java
@@ -1,0 +1,5 @@
+package racingcar.model.car;
+
+public interface Car {
+
+}
