+++ src/test/java/racingcar/model/CarManagerTest.java
@@ -6,6 +6,8 @@ import java.util.List;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.model.car.CarManager;
+import racingcar.model.car.RacingCar;
 
 @DisplayName("[Car Manager]")
 class CarManagerTest {
