+++ src/test/java/racingcar/model/RacingCarTest.java
@@ -11,6 +11,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import racingcar.model.car.RacingCar;
 
 @Nested
 @DisplayName("[Racing Car]")
